Refactor diets and ingredients (hopefully)
