+++ app/models_bak.py
@@ -1,0 +1,17 @@
+# coding: utf-8
+
+import math
+
+import types
+
+import bcrypt
+import hashlib
+
+import unidecode
+
+from flask import current_app as application
+
+from app import db
+from app.auth import login
+
+from sqlalchemy.exc import DatabaseError
