+++ app/templates/diets/edit.html.j2
@@ -20,7 +20,7 @@
     <div class="container">
         <div class="row">
             <div class="col-8">
-                <form action="/diet={{diet.id}}/edit" class="edit__form form-group" method="post" accept-charset="utf-8">
+                <form action="{{ url_for('DietsView:edit', id=diet.id) }}" class="edit__form form-group" method="post" accept-charset="utf-8">
                     <table class="table">
                         <tr>
                             <th scope="col">{{ texts.title }}</th>
@@ -85,8 +85,8 @@
                     <input type="submit" class="btn btn-warning" value="{{ texts.edit_confirm }}" />
                 </form>
 
-                <form action="/diet={{ diet.id }}/remove" method="post" accept-charset="utf-8">
-                    <a href="/diet={{ diet.id }}" class="btn btn-secondary">{{ texts.edit_cancel }} {{ icons.edit }}</a>
+                <form action="{{ url_for('DietsView:delete', id=diet.id) }}" method="post" accept-charset="utf-8">
+                    <a href="{{ url_for('DietsView:show', id=diet.id) }}" class="btn btn-secondary">{{ texts.edit_cancel }} {{ icons.edit }}</a>
                     {% if diet.is_used == False %}
                         <button type="submit" class="btn btn-danger">{{ texts.diet_delete }} {{ icons.delete }}</button>
                     {% else %}
