+++ app/templates/ingredients/all.html.j2
@@ -20,7 +20,7 @@
                 </tr>
                 {% for ingredient in ingredients: %}
                     <tr>
-                        <td><a href="/ingredient={{ingredient.id}}">{{ ingredient.name }}</a></td>
+                        <td><a href="{{ url_for('IngredientsView:show', id=ingredient.id) }}">{{ ingredient.name }}</a></td>
                         <td>{{ ingredient.calorie }}</td>
                         <td>{{ ingredient.protein }}</td>
                         <td>{{ ingredient.fat }}</td>
@@ -28,7 +28,7 @@
                     </tr>
                 {% endfor %}
             </table>
-            <a href="/newingredient" target="_blank"><button class="btn btn-secondary" style="margin-left: 5px">{{ texts.ingredient_add }}</button></a>
+            <a href="{{ url_for('IngredientsView:new') }}" target="_blank"><button class="btn btn-secondary" style="margin-left: 5px">{{ texts.ingredient_add }}</button></a>
         </div> 
     </div>
 {% endblock %}
