+++ app/templates/ingredients/edit.html.j2
@@ -11,7 +11,7 @@
     <div class="container">
         <div class="row">
             <div class="col-md-8 col-sm-12"  style="margin-bottom:30px">
-                <form action="/ingredient={{ingredient.id}}/edit" class="form-group" method="post" accept-charset="utf-8">
+                <form action="{{ url_for('IngredientsView:post', id=ingredient.id) }}" class="form-group" method="post" accept-charset="utf-8">
                     <table class="table">
                         <tr>
                             <th>{{ texts.title }}</th>
@@ -53,15 +53,23 @@
 
 
                         <button type="submit" class="btn btn-warning">{{ texts.edit_confirm }} {{ icons.edit }}</button>
-                        <a class="btn btn-secondary" href="/ingredient={{ ingredient.id }}">{{ texts.edit_cancel }} {{ icons.edit }}</a>
+                        <a class="btn btn-secondary" href="{{ url_for('IngredientsView:show', id=ingredient.id) }}">{{ texts.edit_cancel }} {{ icons.edit }}</a>
                     
-                        {% if ingredient.is_used == True %}
-                            <span class="btn btn-danger" disabled>{{ texts.delete_error }} {{ icons.delete }}</span>
-                        {% else %}
-                            <a class="btn btn-danger" href="/ingredient={{ ingredient.id }}/remove" >{{ texts.ingredient_delete }} {{ icons.delete }}</a>
-                        {% endif %}
 
                 </form>
+
+                <form
+                    action="{{ url_for('IngredientsView:delete', id=ingredient.id) }}"
+                    onsubmit="return confirm('{{ texts.ingredient_delete_confirm }}');"
+                    method="post"
+                    accept-charset="utf-8"
+                >
+                    {% if ingredient.is_used == False %}
+                        <button type="submit" class="btn btn-danger">{{ texts.ingredient_delete }} {{ icons.delete }}</button>
+                    {% else %}
+                        <button type="submit" class="btn btn-danger" disabled>{{ texts.delete_error }} {{ icons.delete }}</button>
+                    {% endif %}
+                 </form>
             </div>
 
              {% include ('recipe/_recipe_list.html.j2') %}
