+++ app/templates/ingredients/show.html.j2
@@ -28,8 +28,8 @@
                     </tr>
                  </table>
 
-                <form action="/ingredient={{ingredient.id}}/remove" onsubmit="return confirm('{{ texts.ingredient_delete_confirm }}');" method="post" accept-charset="utf-8">
-                    <a class="btn btn-secondary" href="/ingredient={{ ingredient.id }}/edit">{{ texts.edit }} {{ icons.edit }}</a>
+                <form action="{{ url_for('IngredientsView:delete', id=ingredient.id) }}" onsubmit="return confirm('{{ texts.ingredient_delete_confirm }}');" method="post" accept-charset="utf-8">
+                    <a class="btn btn-secondary" href="{{ url_for('IngredientsView:edit', id=ingredient.id) }}">{{ texts.edit }} {{ icons.edit }}</a>
                     {% if ingredient.is_used == False %}
                         <button type="submit" class="btn btn-danger">{{ texts.ingredient_delete }} {{ icons.delete }}</button>
                     {% else %}
