+++ ketocalc.py
@@ -3,7 +3,9 @@ from flask import request, redirect
 
 from app import create_app
 
-from app.models import db, User
+# from app.models import db
+# from app import db
+from app.models.users import User
 
 from app.data import template_data
 
@@ -27,6 +29,6 @@ def session_management():
         return redirect("/")
 
 
-@application.shell_context_processor
-def make_shell_context():
-    return {"db": db, "User": User}
+# @application.shell_context_processor
+# def make_shell_context():
+#     return {"db": db, "User": User}
